AI_API_INTERGRATION/requirements.txt
groq
cohere
requests
huggingface_hub
google-generativeai
python-dotenv